requirements.txt
youtube-transcript-api
google-genai